do not assign a name for nodes when they are let-bound
the purpose of this was to make it easier to read the generated LLVM
IR but now this feels problematic to me

as it turned out, varying the metadata of a function results in a new
function that is not = to the previous one. this means assigning a
name to an existing node and then using both the named and the unnamed
node would result in two separate compilations. binding a name to
something should not cause a new compilation of it.